Link function and curve (1/2)
The functions are now connected to their associated curve, thus whenever
the data needs updating, the curve will update itself.
We now require a proper way of creating the curves. The easiest I can
think of would be to create a new curve for every single function. It's
not very heavy since no data needs to be associated with the curve until
display time