Midpoint - working but slow, architecture change needed
The midpoint works if computed for any given value.
But everytime, every single previous point is recomputed, which is a
huge waste.

Instead, an IntegralCurve should be created, that basically does the
same as the FunctionCurve but compute all value.
Alternatively, AbstractFunction could be modified to have a
computeRange(min, max, resolution) function, which would compute all the
data as once.
By redifining it for the IntegralFunction, it would allow to compute it
efficiently.