Fix huge bug with displaying curve on multiple plots
The selection of curves was screwing up which curves are to be
displayed.
Indeed, the same curves can be present in more than one category
(Integral, Hierarchical both linked to Gaussian for instance).
Thus, if one is checked (displayed), the other one can possibly be
unchecked.

For now I've disable the ability to check curves for the integral.
However this isn't an optimal solution.